Désactive les popups également si dessin en cours
